Lots of refactoring in tests. Done with most of mandatory functions.

+++ Libft/ft_strlen.c
@@ -12,9 +12,9 @@
 
 #include "libft.h"
 
-int		ft_strlen(const char *s)
+size_t	ft_strlen(const char *s)
 {
-	int		len;
+	size_t	len;
 
 	len = 0;
 	while (s[len] != '\0')
